Replace pcregrep with awk
This commit updates the logic used to parse system_profiler and
com.apple.Bluetooth data to use awk as opposed to pcrgrep. awk is a
command line tool found on MacOS by default where pcregrep is a tool
which needs to be installed from an external source.

The AirPodsPower.sh script performs and functions the same as before but
this commit should allow for users to more easily install and use the
script.